+++ packages/gateway/src/util/retry.ts
@@ -1,5 +1,5 @@
 async function sleep(ms: number) {
-  return new Promise<void>(resolve => setTimeout(resolve, ms));
+  return new Promise<void>((resolve) => setTimeout(resolve, ms));
 }
 
 export type RetryInput<T> = {
@@ -17,12 +17,12 @@ async function retry<T>({
   maxTries = Number.MAX_SAFE_INTEGER,
   retryExhaustedLog = 'Retry exhausted',
   retryLog,
-  successLog
+  successLog,
 }: RetryInput<T>): Promise<T> {
   let error = new Error(retryExhaustedLog);
 
   for (let tryIdx = 0; tryIdx < maxTries; tryIdx++) {
-    tryIdx == 1 && retryLog && console.log(retryLog);
+    tryIdx > 0 && console.log(retryLog);
 
     try {
       const result = await fn();