+++ packages/plugin-instagram-api/src/middlewares/loginMiddleware.ts
@@ -11,12 +11,12 @@ const loginMiddleware = async (req, res) => {
   const models = await generateModels(subdomain);
   const INSTAGRAM_APP_ID = await getConfig(models, 'INSTAGRAM_APP_ID');
   const INSTAGRAM_APP_SECRET = await getConfig(models, 'INSTAGRAM_APP_SECRET');
+
   const INSTAGRAM_PERMISSIONS = await getConfig(
     models,
     'INSTAGRAM_PERMISSIONS',
-    'pages_messaging,pages_manage_ads,pages_manage_engagement,pages_manage_metadata,pages_read_user_content,instagram_basic,instagram_manage_messages',
+    'pages_messaging,pages_manage_ads,pages_manage_engagement,pages_manage_metadata,pages_read_user_content',
   );
-
   const MAIN_APP_DOMAIN = getEnv({
     name: 'MAIN_APP_DOMAIN',
   });
@@ -28,13 +28,13 @@ const loginMiddleware = async (req, res) => {
   const INSTAGRAM_LOGIN_REDIRECT_URL = await getConfig(
     models,
     'INSTAGRAM_LOGIN_REDIRECT_URL',
-    `${DOMAIN}/gateway/pl:instagram/instagram/login`,
+    `${DOMAIN}/gateway/pl:instagram/iglogin`,
   );
 
   const conf = {
     client_id: INSTAGRAM_APP_ID,
     client_secret: INSTAGRAM_APP_SECRET,
-    scope: INSTAGRAM_PERMISSIONS,
+    scope: `${INSTAGRAM_PERMISSIONS},instagram_basic,instagram_manage_messages,business_management`,
     redirect_uri: INSTAGRAM_LOGIN_REDIRECT_URL,
   };
   debugRequest(debugFacebook, req);
@@ -45,7 +45,7 @@ const loginMiddleware = async (req, res) => {
       client_id: conf.client_id,
       redirect_uri: conf.redirect_uri,
       scope: conf.scope,
-      state: DOMAIN,
+      state: `${DOMAIN}/gateway/pl:instagram`,
     });
 
     // checks whether a user denied the app facebook login/permissions
@@ -103,7 +103,7 @@ const loginMiddleware = async (req, res) => {
       });
 
       for (const integration of integrations) {
-        await repairIntegrations(models, integration.erxesApiId);
+        await repairIntegrations(subdomain, models, integration.erxesApiId);
       }
     } else {
       await models.Accounts.create({